refactor: introduce Main layout and rename features
