+++ src/contents/junit/TestDetails.tsx
